fix: revert exporting test function in sdk
