+++ packages/opentelemetry-core/src/utils/environment.ts
@@ -129,37 +129,3 @@ export function parseEnvironment(values: ENVIRONMENT_MAP): ENVIRONMENT {
 
   return environment;
 }
-
-let lastMock: ENVIRONMENT_MAP = {};
-
-/**
- * Mocks environment used for tests.
- */
-export function mockEnvironment(values: ENVIRONMENT_MAP) {
-  lastMock = values;
-  if (typeof process !== 'undefined') {
-    Object.keys(values).forEach(key => {
-      process.env[key] = String(values[key]);
-    });
-  } else {
-    Object.keys(values).forEach(key => {
-      ((window as unknown) as ENVIRONMENT_MAP)[key] = String(values[key]);
-    });
-  }
-}
-
-/**
- * Removes mocked environment used for tests.
- */
-export function removeMockEnvironment() {
-  if (typeof process !== 'undefined') {
-    Object.keys(lastMock).forEach(key => {
-      delete process.env[key];
-    });
-  } else {
-    Object.keys(lastMock).forEach(key => {
-      delete ((window as unknown) as ENVIRONMENT_MAP)[key];
-    });
-  }
-  lastMock = {};
-}
